[llvm] Move libSupportHTTP to top-level libHTTP (NFC)
