+++ llvm/CMakeLists.txt
@@ -1369,6 +1369,7 @@ endif()
 # Put this before tblgen. Else we have a circular dependence.
 add_subdirectory(lib/Demangle)
 add_subdirectory(lib/Support)
+add_subdirectory(lib/HTTP)
 add_subdirectory(lib/TableGen)
 
 add_subdirectory(utils/TableGen)
