+++ llvm/include/llvm/HTTP/HTTPClient.h
@@ -12,8 +12,8 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#ifndef LLVM_SUPPORT_HTTP_HTTPCLIENT_H
-#define LLVM_SUPPORT_HTTP_HTTPCLIENT_H
+#ifndef LLVM_HTTP_HTTPCLIENT_H
+#define LLVM_HTTP_HTTPCLIENT_H
 
 #include "llvm/ADT/SmallString.h"
 #include "llvm/ADT/SmallVector.h"
@@ -85,4 +85,4 @@ public:
 
 } // end namespace llvm
 
-#endif // LLVM_SUPPORT_HTTP_HTTPCLIENT_H
+#endif // LLVM_HTTP_HTTPCLIENT_H
