+++ llvm/tools/llvm-debuginfod-find/CMakeLists.txt
@@ -2,7 +2,6 @@ set(LLVM_LINK_COMPONENTS
   Option
   Object
   Support
-  SupportHTTP
   )
 set(LLVM_TARGET_DEFINITIONS Opts.td)
 tablegen(LLVM Opts.inc -gen-opt-parser-defs)
@@ -14,6 +13,7 @@ add_llvm_tool(llvm-debuginfod-find
   DebugInfodFindOptsTableGen
   GENERATE_DRIVER
   )
+target_link_libraries(llvm-debuginfod-find PRIVATE LLVMHTTP)
 if(NOT LLVM_TOOL_LLVM_DRIVER_BUILD)
   target_link_libraries(llvm-debuginfod-find PRIVATE LLVMDebuginfod)
 endif()
