+++ llvm/tools/llvm-profdata/CMakeLists.txt
@@ -3,7 +3,6 @@ set(LLVM_LINK_COMPONENTS
   Object
   ProfileData
   Support
-  SupportHTTP
   )
 
 add_llvm_tool(llvm-profdata
@@ -13,4 +12,4 @@ add_llvm_tool(llvm-profdata
   intrinsics_gen
   )
 
-target_link_libraries(llvm-profdata PRIVATE LLVMDebuginfod)
+target_link_libraries(llvm-profdata PRIVATE LLVMHTTP LLVMDebuginfod)
